requirements.txt
# --- Web Framework ---
Flask==3.0.3             # Web server for handling API requests
Werkzeug==3.1.2          # Flask dependency for WSGI utilities

# --- Database and Data Handling ---
SQLAlchemy==2.0.36       # ORM for database interactions
pandas==2.2.3            # Data manipulation library
numpy==1.26.4            # Numerical operations
fsspec==2024.10.0        # Filesystem interface

# --- Machine Learning and Deep Learning ---
torch==2.5.1             # PyTorch for model handling
scikit-learn==1.5.2      # Machine learning algorithms
torchmetrics==0.7.0      # Evaluation metrics for PyTorch
tqdm==4.66.6             # Progress bar for operations
numba==0.60.0            # JIT compiler for optimized operations

# --- Natural Language Processing (NLP) ---
spacy==3.8.2             # NLP library for text processing
spacy-legacy==3.0.12     # Legacy support for spaCy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl  # SpaCy model for text analysis
nltk==3.9.1              # Natural language toolkit for text analysis
llama-index==0.11.22     # LlamaIndex library for indexing
llama-index-agent-openai==0.3.4 # Agent functionality for LlamaIndex
llama-cloud==0.1.4       # Cloud functionality for embedding and retrieval
llama-index-embeddings-openai==0.2.5 # Embeddings for OpenAI
llama-index-llms-openai==0.2.16 # Large language models for OpenAI
llama-index-readers-file==0.2.2 # Llama file reader
llama-parse==0.5.13      # Parsing for LlamaIndex
llama-index-cli==0.3.1   # CLI tools for LlamaIndex
llama-index-program-openai==0.2.0 # Program-based Llama functions

# --- NVIDIA NeMo ---
nemo-toolkit==1.10.0     # NVIDIA NeMo toolkit for large language models

# --- Networking and Async ---
aiohappyeyeballs==2.4.3  # Asynchronous connection handling
aiohttp==3.10.10         # Async HTTP client
aiosignal==1.3.1         # Signal handling for aiohttp
httpx==0.27.2            # HTTP requests with async support
httpcore==1.0.6          # Core for httpx
anyio==4.6.2.post1       # Async I/O framework

# --- Utility and Helpers ---
pydantic==2.9.2          # Data validation and settings management
pydantic_core==2.23.4    # Core library for pydantic
python-dotenv==1.0.1     # Load environment variables
typing-inspect==0.9.0    # Advanced type inspection
typing_extensions==4.12.2 # Extensions for Python typing

# --- Parsing and File Processing ---
PyPDF2==3.0.1            # PDF parsing and manipulation
python-docx==1.1.2       # DOCX file handling
lxml==5.3.0              # XML and HTML parsing library
striprtf==0.0.26         # RTF file handling
Unidecode==1.3.8         # ASCII transliteration of Unicode text

# --- HTTP and Requests ---
requests==2.32.3         # HTTP library
urllib3==2.2.3           # HTTP client for requests
certifi==2024.8.30       # SSL certificates for requests
idna==3.10               # Internationalized Domain Names support

# --- Miscellaneous ---
beautifulsoup4==4.12.3   # HTML/XML parsing library
soupsieve==2.6           # CSS selector support for BeautifulSoup
blinker==1.9.0           # Signal handling in Flask
blis==1.0.1              # spaCy performance enhancement
colorama==0.4.6          # Cross-platform colored terminal text
cymem==2.0.8             # Memory management in spaCy
distro==1.9.0            # Linux distribution information
MarkupSafe==3.0.2        # HTML/XML escaping for Jinja2
jiter==0.7.0             # JIT compiler helper
joblib==1.4.2            # Lightweight pipelining
h11==0.14.0              # HTTP/1.1 for HTTP/2 proxies
rich==13.9.4             # Rich text and formatting
tenacity==8.5.0          # Retry library for Python
click==8.1.7             # Command-line interface creation toolkit
Pygments==2.18.0         # Syntax highlighting
frozendict==2.4.6        # Immutable dictionary
scipy==1.14.1            # Scientific computing library
tzdata==2024.2           # Timezone data
wasabi==1.1.3            # Lightweight logging and utilities


# --- Additional Utilities and Compatibility Libraries ---
smart-open==7.0.5           # File handling for S3, GCP, and Azure, useful in cloud storage scenarios
striprtf==0.0.26            # Parsing library for extracting text from RTF files
future==1.0.0               # Compatibility library for Python 2 and 3
shellingham==1.5.4          # Detects the active shell, useful in command-line interfaces
torchvision                  # PyTorch extension for computer vision tasks (add version if necessary)
torchtext                    # PyTorch extension for text data processing (add version if necessary)
lightning-utilities==0.11.8  # Utility functions for scalable model training, used with PyTorch Lightning
dataclasses-json==0.6.7      # Utility for JSON (de)serialization with Python dataclasses
mypy-extensions==1.0.0       # Extensions for mypy type checker, useful for type safety
